storing extra repayment schedule in the database

+++ database/migrations/2023_08_25_204945_create_extra_repayment_schedules_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
             $table->decimal('interest_component', 10, 2);
             $table->decimal('extra_repayment', 10, 2)->nullable();
             $table->decimal('ending_balance', 10, 2);
-            $table->float('remaining_loan_term');
+            $table->integer('remaining_loan_term');
             $table->foreignId('user_id')->constrained();
             $table->timestamps();
         });
